pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dni-do-emerytury"
version = "0.1.0"
authors = [
  { name="Your Name", email="you@example.com" },
]
description = "Narzędzie CLI do obliczania pozostałych dni i godzin roboczych do emerytury, z uwzględnieniem polskich świąt i konfigurowalnego wieku emerytalnego."
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyYAML",
    "python-dateutil",
    "holidays",
]

[project.scripts]
dni-do-emerytury = "dni_do_emerytury.main:main"

[project.urls]
Homepage = "https://github.com/theundefined/dni-do-emerytury"
Issues = "https://github.com/theundefined/dni-do-emerytury/issues"
